Fetch the API key for authenticated requests
